safest rp 20 Mar adjustments

+++ code/methods/utils/SELD_metrics.py
@@ -12,8 +12,7 @@
 #
 
 import numpy as np
-
-eps = np.finfo(np.float).eps
+eps = np.finfo(float).eps
 from scipy.optimize import linear_sum_assignment
 from IPython import embed
